requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.6.2
charset-normalizer==3.3.2
Django==5.0.6
idna==3.7
pip==24.0
requests==2.32.3
soupsieve==2.5
sqlparse==0.5.0
tzdata==2024.1
urllib3==2.2.1
pillow==10.3.0
django-environ==0.11.2